geoLocation treeshaking updates
