+++ packages/teams-js/src/public/index.ts
@@ -46,13 +46,13 @@ export { EmailAddress } from './emailAddress';
 export * as appInstallDialog from './appInstallDialog';
 export * as barCode from './barCode';
 export * as chat from './chat';
-//It is necessary to export the OpenGroupChatRequest and OpenSingleChatRequest from chat.ts individually as well
-//to keep the named exports so as to not break the existing consumers using the named exports.
+//It is necessary to export the OpenGroupChatRequest and OpenSingleChatRequest interfaces from chat.ts individually as well
+//to keep the named exports so as to not break the existing consumers directly referencing the named exports.
 export { OpenGroupChatRequest, OpenSingleChatRequest } from './chat';
 export * as clipboard from './clipboard';
 export { dialog } from './dialog';
 export { nestedAppAuth } from './nestedAppAuth';
-export { geoLocation } from './geoLocation';
+export * as geoLocation from './geolocation/geoLocation';
 export { getAdaptiveCardSchemaVersion } from './adaptiveCards';
 export { pages } from './pages';
 export {
